Make reporter format fully configurable
This commit extracts the formatting logic from the reporter, so that it
can easily be overridden and adapted to suit the needs of everybody.

The current format is now called `DefaultReporterFormat`. It is also
slightly modified, so that the line number is shown only once.

Implementing a new format is simple:

  reporterConfig := reporterConfig.value.withFormat(reporter =>
    new ReporterFormat(reporter) {
      override def formatProblem(problem: Problem): String = // ...
      override def printSummary(): Unit = // ...
    })